Move TopographyOptions to selection module
